Changed date thingy, guest thingy, free rooms thingy.
Changed so you an arrive on some1s departure date
Changed so guests has a status field were you can see if they are arriving, leaving or checked in.
Changed so you get a list of rooms to choose from when checking for available rooms
Fixed some Calendar date things that was annoying. 
Added one or two tests