Created logic class for gui and dateLogic class.
Added GuiLogic class that handle a lot of the code that was previously in the frames.
Added a dateLogic class that takes care of converting some dates and getting a current time without hours, seconds and minutes.
Changed landingPage to have set currentUser in guiLogic.
Pulled down load of code from the frame into the guiLogic. A lot of them being filling of Table models.
Changed roomInfo frame so it also adds rows instead of creating a whole new defaultTableModel, so I could pull the code down to guiLogic.
GuiLogic also allows for some code reuse and if created correctly possibly more code reuse.