Implemented generalSearch and more abstraction in all mappers.
Added tableName and classType in abstractMapper and a constructor for it.
Added 2 version of generalSearch one specific for the search function inside the mappers and one to use by other mapper methods.
Added a sequence Number method as the new version of the execture and gather doesn't work with none fully mapped classes.
Added tests and made sure as much as I could that it all works.